PainterExamples GCC ambiguity fix
git-svn-id: svn://ultimatepp.org/upp/trunk@934 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ examples/PainterExamples/Big.cpp
@@ -6,8 +6,8 @@ void Big(Painter& sw)
 {
 	int n = 0;
 	double sgn = 1;
-	for(int r = 400; r > 5; r -= 3) {
-		for(int i = 0; i < 400; i++) {
+	for(double r = 400; r > 5; r -= 3) {
+		for(double i = 0; i < 400; i++) {
 			Pointf p = Polar(sgn * i * M_2PI / 400) * r + Pointf(400, 400);
 			if(i)
 				sw.Line(p);
